GitButler Integration Commit
This is an integration commit for the virtual branches that GitButler is tracking.

Due to GitButler managing multiple virtual branches, you cannot switch back and
forth between git branches and virtual branches easily. 

If you switch to another branch, GitButler will need to be reinitialized.
If you commit on this branch, GitButler will throw it away.

Here are the branches that are currently applied:
 - Add status component and variants (refs/gitbutler/Add-status-component-and-variants)
   - src/ui/Alert/Alert.stories.tsx
   - src/ui/Button/Button.variants.ts
   - src/shared/classes.ts
   - src/layout/Wrapper/Wrapper.variants.ts
   - src/ui/Plan/Plan.tsx
   - src/ui/Modal/Modal.variants.ts
   - src/ui/Plan/Plan.variants.ts
   - tailwind.config.ts
   - src/styles/tailwind.css
   - src/index.ts
   - src/ui/Plan/Plan.stories.tsx
   - src/ui/Alert/Alert.tsx
   - src/ui/Plan/index.ts
 - Add tailwind and Chromatic CI
 (refs/gitbutler/Add-tailwind-and-Chromatic-CI-)
   branch head: bb864771e1c4b08c39d6404ff65c64d6f0a4edaa

Your previous branch was: refs/heads/main

The sha for that commit was: 4638343c915712829bfdab75204149ba6a8400e2

For more information about what we're doing here, check out our docs:
https://docs.gitbutler.com/features/virtual-branches/integration-branch